SubGhz: add CC112x driver

+++ firmware/targets/furi_hal_include/furi_hal_subghz.h
@@ -78,7 +78,7 @@ void furi_hal_subghz_load_preset(FuriHalSubGhzPreset preset);
  *
  * @param      data  Registers data
  */
-void furi_hal_subghz_load_registers(const uint8_t data[][2]);
+void furi_hal_subghz_load_registers(const uint16_t data[][2]);
 
 /** Load PATABLE
  *
